Fix up refactoring
